+++ Cargo.toml
@@ -26,3 +26,4 @@ tracing-subscriber = { version = "0.3" }
 winter-maybe-async = "0.10"
 hex = "0.4.3"
 rand_chacha = "0.3"
+once_cell = "1.20"
